backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.0

# Production Server
gunicorn>=21.0.0
whitenoise>=6.6.0

# Environment
python-dotenv>=1.0.0

# AI / RAG
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
numpy>=1.24.0

# Para migración futura a pgvector:
# pgvector>=0.2.0
